fix: resolve macOS build warnings from Qt 6.10.3 upgrade
- Use qsizetype for indexOf/size() results in usbdecentscale, shotserver,
  serialtransport, screencaptureservice to fix -Wshorten-64-to-32
- Use static_cast<int> for removeFavoriteProfile calls in profilemanager
  where the API takes int but the loop index is qsizetype
- Bump macOS deployment target 12.0 → 13.0: Qt 6.10.3 permission plugins
  (Bluetooth, Location) are built for macOS 13.0 minimum

+++ src/controllers/profilemanager.cpp
@@ -634,7 +634,7 @@ bool ProfileManager::deleteProfile(const QString& filename) {
             QVariantList favorites = m_settings->favoriteProfiles();
             for (qsizetype i = 0; i < favorites.size(); ++i) {
                 if (favorites[i].toMap()["filename"].toString() == filename) {
-                    m_settings->removeFavoriteProfile(i);
+                    m_settings->removeFavoriteProfile(static_cast<int>(i));
                     break;
                 }
             }
@@ -1171,7 +1171,7 @@ void ProfileManager::refreshProfiles() {
             QString fn = favorites.at(i).toMap()[QStringLiteral("filename")].toString();
             if (!known.contains(fn)) {
                 qWarning() << "refreshProfiles: removing stale favorite" << fn << "(profile not found)";
-                m_settings->removeFavoriteProfile(i);
+                m_settings->removeFavoriteProfile(static_cast<int>(i));
             }
         }
